minor patch fix
line-by-line documentation for most of the interface elements according to screenshots of the main blocks.
the ability to artificially control the speed of the Internet connection at the browser level to stabilize the loading of interface elements.
most of the manual delays in the program have been changed from per-second to pixel verification.
mouse is again assigned as the main peripheral input device (due to increased stability after entering all the above items), keyboard is secondary.

test possibility of working with the queue interface (elimination of directions) through the diagnostic workstation.
test possibility of working with the interface of the directions journal.

other minor changes.